requirements.txt
beautifulsoup4      4.8.2
certifi             2019.11.28
cffi                1.14.0
covid               2.1.4
cryptography        2.8
decorator           4.4.2
future              0.18.2
pip                 20.0.2
pycparser           2.20
pydantic            1.4
pyTelegramBotAPI    3.6.7
python-telegram-bot 12.5.1
requests            2.7.0
setuptools          41.2.0
six                 1.14.0
soupsieve           2.0
telebot             0.0.3
tornado             6.0.4